py/instalog/requirements.txt
google-auth==1.21.1
google-cloud-bigquery==1.27.2
google-cloud-storage==1.31.0
jsonrpclib-pelix==0.4.1
python-gnupg==0.4.3
PyYAML==5.3.1
requests==2.24.0
